+++ stisim/interventions/base_interventions.py
@@ -7,10 +7,7 @@ import numpy as np
 import sciris as sc
 import pandas as pd
 from scipy.interpolate import interp1d
-
-
-# %% Helper functions
-def count(arr): return np.count_nonzero(arr)
+from stisim.utils import count
 
 
 # %% Base classes
@@ -307,7 +304,7 @@ class SymptomaticTesting(STITest):
     def init_pre(self, sim):
         super().init_pre(sim)
         if self.treat_prob_data is not None:
-            self.treat_prob = np.interp(sim.yearvec, self.treat_prob_data.year.values, self.treat_prob_data.treat_prob.values)
+            self.treat_prob = sc.smoothinterp(sim.yearvec, self.treat_prob_data.year.values, self.treat_prob_data.treat_prob.values, smoothness=0)
         return
 
     def init_results(self):